[refactor] Move util/navDictionary under utils/
