+++ src/content/nav/ja.ts
@@ -1,4 +1,4 @@
-import { navDictionary } from "../../util/navDictionary";
+import { navDictionary } from "../../utils/navDictionary.ts";
 
 /**
  * Japanese labels for navigation items
